Add files via upload
This code is written in Pseudocode from the 9618 CIE AS & A Level Computer Science syllabus. It helps learn and practice stuff to do with procedures and functions from that syllabus. To run the code, run it with the editor in the website below. It goes in a lot of detail.

https://pseudocode.pro/editor